fix(functional-tests): failures due to rate limiting in coupon tests Because:
* coupon tests are failing in stage due to IP rate limiting

This commit:

* adds signin as a pre-requisite to coupon tests (since all subscription tests now require signin, the steps have been encapsulated in an automatic fixture to reduce code duplication)

Closes # FXA-10045
