+++ packages/functional-tests/tests/subscription-tests/coupons-ui-tests/couponForeverDiscount.spec.ts
@@ -2,34 +2,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Page, expect, test } from '../../../lib/fixtures/standard';
 import { VALID_VISA } from '../../../lib/paymentArtifacts';
-import { BaseTarget, Credentials } from '../../../lib/targets/base';
-import { TestAccountTracker } from '../../../lib/testAccountTracker';
 import { Coupon } from '../../../pages/products';
-import { SettingsPage } from '../../../pages/settings';
-import { SigninPage } from '../../../pages/signin';
+import { expect, test } from '../subscriptionFixtures';
 
 test.describe('severity-2 #smoke', () => {
   test.describe('coupon test forever discount', () => {
     test('subscribe successfully with a forever discount coupon', async ({
-      target,
       page,
-      pages: { relier, settings, signin, subscribe },
-      testAccountTracker,
+      pages: { relier, signin, subscribe },
+      credentials,
     }, { project }) => {
       test.skip(
         project.name === 'production',
         'no real payment method available in prod'
       );
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
       await relier.goto();
       await relier.clickSubscribe6Month();
 
@@ -62,23 +49,14 @@ test.describe('severity-2 #smoke', () => {
     });
 
     test('subscribe with credit card and use coupon', async ({
-      target,
       page,
-      pages: { relier, settings, signin, subscribe },
-      testAccountTracker,
+      pages: { relier, signin, subscribe },
+      credentials,
     }, { project }) => {
       test.skip(
         project.name === 'production',
         'no real payment method available in prod'
       );
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
       await relier.goto();
       await relier.clickSubscribe6Month();
 
@@ -104,21 +82,3 @@ test.describe('severity-2 #smoke', () => {
     });
   });
 });
-
-async function signInAccount(
-  target: BaseTarget,
-  page: Page,
-  settings: SettingsPage,
-  signin: SigninPage,
-  testAccountTracker: TestAccountTracker
-): Promise<Credentials> {
-  const credentials = await testAccountTracker.signUp();
-  await page.goto(target.contentServerUrl);
-  await signin.fillOutEmailFirstForm(credentials.email);
-  await signin.fillOutPasswordForm(credentials.password);
-
-  await expect(page).toHaveURL(/settings/);
-  await expect(settings.settingsHeading).toBeVisible();
-
-  return credentials;
-}
