+++ packages/functional-tests/tests/subscription-tests/coupons-ui-tests/paymentMode.spec.ts
@@ -2,14 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Page, expect, test } from '../../../lib/fixtures/standard';
 import { VALID_MASTERCARD, VALID_VISA } from '../../../lib/paymentArtifacts';
-import { BaseTarget, Credentials } from '../../../lib/targets/base';
-import { TestAccountTracker } from '../../../lib/testAccountTracker';
 import { Coupon } from '../../../pages/products';
 import { SubscriptionManagementPage } from '../../../pages/products/subscriptionManagement';
-import { SettingsPage } from '../../../pages/settings';
-import { SigninPage } from '../../../pages/signin';
+import { expect, test } from '../subscriptionFixtures';
 
 test.describe('severity-2 #smoke', () => {
   test.describe('payment', () => {
@@ -17,20 +13,12 @@ test.describe('severity-2 #smoke', () => {
       target,
       page,
       pages: { relier, subscribe, settings, signin },
-      testAccountTracker,
+      credentials,
     }, { project }) => {
       test.skip(
         project.name === 'production',
         'no real payment method available in prod'
       );
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
       await relier.goto();
       await relier.clickSubscribe6Month();
 
@@ -71,21 +59,3 @@ test.describe('severity-2 #smoke', () => {
     });
   });
 });
-
-async function signInAccount(
-  target: BaseTarget,
-  page: Page,
-  settings: SettingsPage,
-  signin: SigninPage,
-  testAccountTracker: TestAccountTracker
-): Promise<Credentials> {
-  const credentials = await testAccountTracker.signUp();
-  await page.goto(target.contentServerUrl);
-  await signin.fillOutEmailFirstForm(credentials.email);
-  await signin.fillOutPasswordForm(credentials.password);
-
-  await expect(page).toHaveURL(/settings/);
-  await expect(settings.settingsHeading).toBeVisible();
-
-  return credentials;
-}
