+++ packages/functional-tests/tests/subscription-tests/support.spec.ts
@@ -2,18 +2,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { SettingsPage } from '../..//pages/settings';
-import { Page, expect, test } from '../../lib/fixtures/standard';
 import { VALID_VISA } from '../../lib/paymentArtifacts';
-import { BaseTarget, Credentials } from '../../lib/targets/base';
-import { TestAccountTracker } from '../../lib/testAccountTracker';
 import { SubscriptionManagementPage } from '../../pages/products/subscriptionManagement';
 import {
   App,
   SubscriptionSupportPage,
   Topic,
 } from '../../pages/products/support';
-import { SigninPage } from '../../pages/signin';
+import { expect, test } from './subscriptionFixtures';
 
 test.describe('severity-1 #smoke', () => {
   test.describe('support form without valid session', () => {
@@ -31,12 +27,10 @@ test.describe('severity-1 #smoke', () => {
 
   test.describe('support form without active subscriptions', () => {
     test('go to support form, redirects to subscription management, then back to settings', async ({
-      page,
       target,
-      pages: { settings, signin },
-      testAccountTracker,
+      page,
+      pages: { settings },
     }) => {
-      await signInAccount(target, page, settings, signin, testAccountTracker);
       await page.goto(`${target.contentServerUrl}/support`);
 
       await expect(page).toHaveURL(/settings/);
@@ -54,21 +48,12 @@ test.describe('severity-2 #smoke', () => {
       target,
       page,
       pages: { relier, subscribe, settings, signin },
-      testAccountTracker,
+      credentials,
     }, { project }) => {
       test.skip(
         project.name === 'production',
         'no real payment method available in prod'
       );
-
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
       await relier.goto();
       await relier.clickSubscribe();
 
@@ -112,21 +97,3 @@ test.describe('severity-2 #smoke', () => {
     });
   });
 });
-
-async function signInAccount(
-  target: BaseTarget,
-  page: Page,
-  settings: SettingsPage,
-  signin: SigninPage,
-  testAccountTracker: TestAccountTracker
-): Promise<Credentials> {
-  const credentials = await testAccountTracker.signUp();
-  await page.goto(target.contentServerUrl);
-  await signin.fillOutEmailFirstForm(credentials.email);
-  await signin.fillOutPasswordForm(credentials.password);
-
-  await expect(page).toHaveURL(/settings/);
-  await expect(settings.settingsHeading).toBeVisible();
-
-  return credentials;
-}
